+++ src/Presentation/SmartStore.Web/Models/Catalog/ProductDetailsModel.cs
@@ -200,8 +200,9 @@ namespace SmartStore.Web.Models.Catalog
 			public int ProductId { get; set; }
 
 			public bool HidePrices { get; set; }
+            public bool ShowLoginNote { get; set; }
 
-			public bool DynamicPriceUpdate { get; set; }
+            public bool DynamicPriceUpdate { get; set; }
 			public bool BundleItemShowBasePrice { get; set; }
 
 			public string NoteWithDiscount { get; set; }